+++ routes/index.js
@@ -1,6 +1,6 @@
 module.exports = {
   api: require('./api'),
+  wmts: require('./wmts'),
   mbtiles: require('./mbtiles'),
-  permissions: require('./permissions'),
-  wmts: require('./wmts')
+  permissions: require('./permissions')
 }