+++ routes/mbtiles.js
@@ -1,5 +1,5 @@
 const fs = require('fs')
-const Conf = require('conf')
+const Conf = require('../plugins/conf')
 const path = require('path')
 const router = require('express').Router()
 const MBTiles = require('mbtiles-offline')